// EcoDOC/2.Source/1.Development/2.Frontend/src/app/core/model/domain/preferences.ts
export class Preferences {
    name: string;
    active: boolean;
    // create_by: number;
    // create_date: Date;
    // update_by: number;
    // update_time: Date;
    // client_id: number;
    orgId: number;
    extension: string;
    maxSize: number;
    preType: string;
    constructor(){
        
    }
}